+++ application/infrastructure/chat.py
@@ -1,4 +1,4 @@
-# app/infrastructure/chat.py
+# application/infrastructure/chat.py
 
 from flask import jsonify
 from db import db
@@ -8,7 +8,7 @@ from flask_jwt_extended import create_access_token, jwt_required, \
     get_jwt_identity, \
     unset_jwt_cookies
 from dotenv import load_dotenv
-from app.infrastructure.message import Message
+from application.infrastructure.message import Message
 load_dotenv()
 
 # Modèle d'un chat
